Refactor: move functions to new component
